+++ wire/cli/default.nix
@@ -1,11 +1,25 @@
+{ getSystem, inputs, ... }:
 {
   perSystem =
     {
       pkgs,
+      lib,
       self',
       buildRustProgram,
+      system,
       ...
     }:
+    let
+      postBuild = ''
+        wrapProgram $out/bin/wire \
+                    --set WIRE_RUNTIME ${../../runtime} \
+                    --set WIRE_KEY_AGENT ${self'.packages.agent} \
+      '';
+      cleanSystem = system: lib.replaceStrings [ "-" ] [ "_" ] system;
+      agents = lib.strings.concatMapStrings (
+        system: "--set WIRE_KEY_AGENT_${cleanSystem system} ${(getSystem system).packages.agent} "
+      ) (import inputs.linux-systems);
+    in
     {
       packages = {
         default = self'.packages.wire;
@@ -30,9 +44,19 @@
             pkgs.makeWrapper
           ];
           postBuild = ''
-            wrapProgram $out/bin/wire \
-                --set WIRE_RUNTIME ${../../runtime} \
-                --set WIRE_KEY_AGENT ${self'.packages.agent}
+            ${postBuild} ${agents}
+          '';
+          meta.mainProgram = "wire";
+        };
+
+        wire-small = pkgs.symlinkJoin {
+          name = "wire";
+          paths = [ self'.packages.wire-unwrapped ];
+          nativeBuildInputs = [
+            pkgs.makeWrapper
+          ];
+          postBuild = ''
+            ${postBuild} --set WIRE_KEY_AGENT_${cleanSystem system} ${self'.packages.agent}
           '';
           meta.mainProgram = "wire";
         };