+++ flake.nix
@@ -12,6 +12,10 @@
 
     # determines systems available for deployment
     linux-systems.url = "github:nix-systems/default-linux";
+
+    # testing inputs
+    nixpkgs_current_stable.url = "github:NixOS/nixpkgs/nixos-25.05";
+    nixpkgs_prev_stable.url = "github:NixOS/nixpkgs/nixos-24.11";
   };
   outputs =
     {