+++ tests/nix/suite/test_remote_deploy/hive.nix
@@ -1,5 +1,5 @@
 let
-  mkHiveNode = import ../utils.nix { testName = "test_remote_deploy"; };
+  mkHiveNode = import ../utils.nix { testName = "test_remote_deploy-@IDENT@"; };
 in
 {
   meta.nixpkgs = import <nixpkgs> { system = "x86_64-linux"; };