+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+## [0.3.0] - 2025-06-20
+
+### Added
+
 - Run tests against `unstable` and `25.05` by @mrshmllow in https://github.com/wires-org/wire/pull/176.
 
 ### Changed